Keycode refactor
